Change to use PointerEvent instead of MouseEvent
The use of `requestPointerLock` has a downside that triggers "Pointer is locked" warning on browser. They are changed to `setPointerCapture`.

`setPointerCapture` has a downside that don't keep the position of pointer when dragged. However, the benefit of `PointerEvent` that can handle touch interactions outweighs this downside.

This change also includes a fix that WaveView draws dashed lines on Safari. Search this diff with `setLineDash([])` to see the change.